… Gemfile.lock
… mise.toml