### Require
* https://github.com/ontoportal-lirmm/ontologies_api_ruby_client/pull/25 
### Context
before 7 API calls, total 7136ms
```log
Processing by ConceptsController#index as HTML
  Parameters: {"conceptid"=>"http://aims.fao.org/aos/agrovoc/c_330834", "concept_schemes"=>"", "language"=>"EN", "ontology"=>"AGROVOC"}
* Locale set to 'en'
Getting: https://data.stageportal.lirmm.fr with {} (t: 0.2040101170005073s - cache: miss)
Getting: https://data.stageportal.lirmm.fr/slices with {:display_links=>false, :display_context=>false} (t: 2.345375471999432s - cache: miss)
Getting: https://data.stageportal.lirmm.fr with {} (t: 0.22914239899910172s - cache: miss)
Getting: https://data.stageportal.lirmm.fr/ontologies/AGROVOC with {:include_views=>true} (t: 0.1970192330009013s - cache: miss)
Getting: https://data.stageportal.lirmm.fr/ontologies/AGROVOC/latest_submission with {:include=>"uriRegexPattern,preferredNamespaceUri"} (t: 0.34039248199951544s - cache: miss)
Getting: https://data.stageportal.lirmm.fr/ontologies/AGROVOC/classes/http%3A%2F%2Faims.fao.org%2Faos%2Fagrovoc%2Fc_330834 with {:dispay=>"prefLabel", :include=>"prefLabel,definition,synonym,obsolete,hasChildren,inScheme,memberOf"} (t: 0.3808957980018022s - cache: miss)
Getting: https://data.stageportal.lirmm.fr/ontologies/AGROVOC/classes/http%3A%2F%2Faims.fao.org%2Faos%2Fagrovoc%2Fc_330834/children with {:pagesize=>750, :language=>"EN", :display=>"prefLabel,obsolete,hasChildren", :include=>"prefLabel,definition,synonym,obsolete,hasChildren,inScheme,memberOf"} (t: 2.8401889389988355s - cache: miss)
Completed 200 OK in 7136ms (Views: 0.3ms | ActiveRecord: 0.0ms | Allocations: 394551)
```
after  3 API calls, total 3975ms
```log
Processing by ConceptsController#index as HTML
  Parameters: {"conceptid"=>"http://aims.fao.org/aos/agrovoc/c_330834", "concept_schemes"=>"", "language"=>"EN", "ontology"=>"AGROVOC"}
* Locale set to 'en'
Getting: https://data.stageportal.lirmm.fr/slices with {:display_links=>false, :display_context=>false} (t: 0.37758208299055696s - cache: miss)
Getting: https://data.stageportal.lirmm.fr/ontologies/AGROVOC/latest_submission with {:include=>"uriRegexPattern,preferredNamespaceUri"} (t: 0.451797458998044s - cache: miss)
Getting: https://data.stageportal.lirmm.fr/ontologies/AGROVOC/classes/http%3A%2F%2Faims.fao.org%2Faos%2Fagrovoc%2Fc_330834/children with {:pagesize=>750, :language=>"EN", :display=>"prefLabel,obsolete,hasChildren"} (t: 2.3704248760041082s - cache: miss)
Completed 200 OK in 3623ms (Views: 0.1ms | ActiveRecord: 0.0ms | Allocations: 301040)
```