@Bilelkihal 2025-02-04
@syphax-bouazzouni 
I couldn't run locally cause bundler 2.5.22 is required somewhere ! (we can't use this version of bundler before going with ruby 3)
do you know what could be the potential reason for this?
![image](https://github.com/user-attachments/assets/58342af2-4138-40c9-82b7-ae8a28a52d62)