@syphax-bouazzouni 2025-02-04
> @syphax-bouazzouni I couldn't run locally cause bundler 2.5.22 is required somewhere ! (we can't use this version of bundler before going with ruby 3) do you know what could be the potential reason for this? ![image](https://private-user-images.githubusercontent.com/61744974/409563255-58342af2-4138-40c9-82b7-ae8a28a52d62.png?jwt=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiJnaXRodWIuY29tIiwiYXVkIjoicmF3LmdpdGh1YnVzZXJjb250ZW50LmNvbSIsImtleSI6ImtleTUiLCJleHAiOjE3Mzg2NzgwNTAsIm5iZiI6MTczODY3Nzc1MCwicGF0aCI6Ii82MTc0NDk3NC80MDk1NjMyNTUtNTgzNDJhZjItNDEzOC00MGM5LTgyYjctYWU4YTI4YTUyZDYyLnBuZz9YLUFtei1BbGdvcml0aG09QVdTNC1ITUFDLVNIQTI1NiZYLUFtei1DcmVkZW50aWFsPUFLSUFWQ09EWUxTQTUzUFFLNFpBJTJGMjAyNTAyMDQlMkZ1cy1lYXN0LTElMkZzMyUyRmF3czRfcmVxdWVzdCZYLUFtei1EYXRlPTIwMjUwMjA0VDE0MDIzMFomWC1BbXotRXhwaXJlcz0zMDAmWC1BbXotU2lnbmF0dXJlPTBlMDk2MTM1NWQ3MWQ4ZmMxNjMxNTZhMjkyMWRjNjRmMjg4ZmRkMTU3YzJlNjE4OTEwZjNmZmQyMGRiNGM1NDUmWC1BbXotU2lnbmVkSGVhZGVycz1ob3N0In0.DTpReMe0xBEk1uq-N_i_FY9jd0dE0_QEUP6oFoXYJWY)

You can do bundle update locally in your machine, to use ruby 2.7.8, in my case I had ruby 3.