@syphax-bouazzouni 2025-02-04
> I got this error when I tried to run it on docker, and doing bundle update also gives the same error, and doing bundle update locally didn't work for me with another error. can you do bundle update from your side and push the gem.lock file since the tests are not working

Done